requirements.txt
Pillow
requests
notebook
jupyter
